Fix compilation errors in C++17

+++ src/entropy/RiceGolombDecoder.hpp
@@ -47,22 +47,22 @@ namespace kanzi
 
    inline byte RiceGolombDecoder::decodeByte()
    {
-       uint64 q = 0;
+       int q = 0;
 
        // quotient is unary encoded
        while (_bitstream.readBit() == 0)
           q++;
 
        // remainder is binary encoded
-       const byte res = byte((q << _logBase) | _bitstream.readBits(_logBase));
+       const int res = (q << _logBase) | int(_bitstream.readBits(_logBase));
 
        if ((_signed == true) && (res != 0))
        {
           if (_bitstream.readBit() == 1)
-             return -res;
+             return byte(-res);
        }
 
-       return res;
+       return byte(res);
    }
 }
 #endif
